@@ -2,32 +2,40 @@
   "name": "todo-list",
   "version": "1.0.0",
   "description": "Task planner for The Odin Project",
+  "type": "module",
   "private": true,
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "webpack --mode=production",
     "watch": "webpack --watch",
-    "start": "webpack serve --open",
+    "start": "webpack serve",
+    "lint": "eslint --ext .js src",
+    "fix": "eslint --ext .js src --fix",
+    "pretty": "prettier --write src",
     "push": "git subtree push --prefix dist origin gh-pages"
   },
   "keywords": [],
   "author": "ChiefWoods",
   "license": "ISC",
   "devDependencies": {
-    "css-loader": "^6.8.1",
-    "eslint": "^8.50.0",
-    "eslint-config-prettier": "^9.0.0",
-    "eslint-config-standard": "^17.0.0",
-    "eslint-plugin-import": "^2.27.5",
+    "css-loader": "^6.10.0",
+    "esbuild": "^0.20.2",
+    "esbuild-loader": "^4.1.0",
+    "eslint": "^8.57.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-config-standard": "^17.1.0",
+    "eslint-plugin-import": "^2.29.1",
     "eslint-plugin-n": "^15.7.0",
     "eslint-plugin-promise": "^6.1.1",
-    "html-webpack-plugin": "^5.5.0",
-    "style-loader": "^3.3.2",
-    "webpack": "^5.78.0",
+    "html-webpack-plugin": "^5.6.0",
+    "prettier": "^3.2.5",
+    "style-loader": "^3.3.4",
+    "webpack": "^5.91.0",
     "webpack-cli": "^5.1.4",
-    "webpack-dev-server": "^4.15.1"
+    "webpack-dev-server": "^4.15.2"
   },
   "dependencies": {
-    "date-fns": "^2.30.0"
+    "date-fns": "^3.6.0"
   }
 }
