@@ -1,43 +1,73 @@
-const path = require('path');
-const HtmlWebpackPlugin = require('html-webpack-plugin');
+import HtmlWebpackPlugin from "html-webpack-plugin";
+import { EsbuildPlugin } from "esbuild-loader";
+import esbuild from "esbuild";
 
-module.exports = {
-  mode: 'production',
-  entry: './src/index.js',
-  devtool: 'inline-source-map',
+export default {
+  mode: "development",
+  entry: "./src/index.js",
+  devtool: "eval-source-map",
   devServer: {
-    static: './dist'
+    static: "dist",
+    client: {
+      logging: "error",
+      overlay: false,
+    },
   },
   plugins: [
     new HtmlWebpackPlugin({
-      title: 'Todo List',
-      favicon: './src/icons/clipboard-check-outline.svg',
-      template: './src/template.html'
-    })
+      title: "Todo List",
+      template: "src/template.html",
+      favicon: "src/icons/clipboard-check-outline.svg",
+    }),
   ],
   output: {
-    filename: 'main.js',
-    path: path.resolve(__dirname, 'dist'),
-    assetModuleFilename: pathData => {
-      const filepath = path.dirname(pathData.filename).split('/').slice(1).join('/');
-      return `${filepath}/[name][ext]`;
-    },
-    clean: true
+    publicPath: "auto",
+    filename: "main.[contenthash].js",
+    assetModuleFilename: "[path]/[name].[contenthash].[ext]",
+    clean: true,
   },
   module: {
     rules: [
       {
         test: /\.css$/i,
-        use: ['style-loader', 'css-loader']
+        use: [
+          "style-loader",
+          "css-loader",
+          {
+            loader: "esbuild-loader",
+            options: {
+              minify: true,
+              implementation: esbuild,
+            },
+          },
+        ],
       },
       {
         test: /\.(png|svg|jpg|jpeg|gif)$/i,
-        type: 'asset/resource'
+        type: "asset/resource",
       },
       {
         test: /\.(woff|woff2|eot|ttf|otf)$/i,
-        type: 'asset/resource'
-      }
-    ]
-  }
+        type: "asset/resource",
+      },
+      {
+        test: /.js$/,
+        loader: "esbuild-loader",
+        options: {
+          target: "esnext",
+          implementation: esbuild,
+        },
+      },
+    ],
+  },
+  optimization: {
+    minimizer: [
+      new EsbuildPlugin({
+        target: "esnext",
+        implementation: esbuild,
+        css: true,
+      }),
+    ],
+  },
+  stats: 'minimal',
 };
